Extend polling timeout
PR #342 had an unexpected side effect where `IDLE_TIMELAPSE` started being heeded correctly. So all event polling ran for at most 60s. This is way too little for the offchain worker (that needs to wait possibly for days between jobs), and even for the offchain compute client (that needs to wait for the job to complete).

This PR extends the timeout to the point where it is practically disabled, but given that that is how we operated this whole time, the change should be non-controversial.

Should we do the same in Go?

More context: #842